+++ Sources/StreamChatSwiftUI/ChatChannel/ChannelInfo/ChatChannelInfoView.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2025 Stream.io Inc. All rights reserved.
+// Copyright © 2026 Stream.io Inc. All rights reserved.
 //
 
 import StreamChat
@@ -161,7 +161,7 @@ struct ChatChannelInfoViewHeaderViewModifier: ViewModifier {
     @Injected(\.colors) private var colors
     @Injected(\.fonts) private var fonts
     
-    let viewModel: ChatChannelInfoViewModel
+    @ObservedObject var viewModel: ChatChannelInfoViewModel
     
     func body(content: Content) -> some View {
         if #available(iOS 26.0, *) {