+++ StreamChatSwiftUITestsAppTests/Pages/ThreadPage.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2025 Stream.io Inc. All rights reserved.
+// Copyright © 2026 Stream.io Inc. All rights reserved.
 //
 
 import Foundation
@@ -7,4 +7,5 @@ import XCTest
 
 class ThreadPage: MessageListPage {
     static var alsoSendInChannelCheckbox: XCUIElement { app.buttons["SendInChannelView"] }
+    static var repliesCountLabel: XCUIElement { app.staticTexts["textLabel"] }
 }