+++ StreamChatSwiftUITestsAppTests/Tests/Ephemeral_Messages_Tests.swift
@@ -1,15 +1,10 @@
 //
-// Copyright © 2025 Stream.io Inc. All rights reserved.
+// Copyright © 2026 Stream.io Inc. All rights reserved.
 //
 
 import XCTest
 
 final class Ephemeral_Messages_Tests: StreamTestCase {
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        assertMockServer()
-    }
-
     func test_userObservesAnimatedGiphy_whenUserAddsGiphyMessage() throws {
         linkToScenario(withId: 435)
 
@@ -19,7 +14,7 @@ final class Ephemeral_Messages_Tests: StreamTestCase {
                 .openChannel()
         }
         WHEN("user sends a giphy using giphy command") {
-            userRobot.sendGiphy(useComposerCommand: true)
+            userRobot.uploadGiphy(useComposerCommand: true)
         }
         THEN("user observes the animated gif") {
             userRobot.assertGiphyImage()
@@ -35,7 +30,7 @@ final class Ephemeral_Messages_Tests: StreamTestCase {
                 .openChannel()
         }
         WHEN("participant sends a giphy") {
-            participantRobot.sendGiphy()
+            participantRobot.uploadGiphy()
         }
         THEN("user observes the animated gif") {
             userRobot.assertGiphyImage()
@@ -74,15 +69,13 @@ final class Ephemeral_Messages_Tests: StreamTestCase {
     func test_channelListNotModified_whenEphemeralMessageShown() throws {
         linkToScenario(withId: 438)
 
-        throw XCTSkip("Check out SWUI-252")
-
         GIVEN("user opens a channel") {
             userRobot
                 .login()
                 .openChannel()
         }
         WHEN("user runs a giphy command") {
-            userRobot.sendGiphy(send: false)
+            userRobot.uploadGiphy(send: false)
         }
         WHEN("user goes back to channel list") {
             userRobot.tapOnBackButton()
@@ -94,8 +87,8 @@ final class Ephemeral_Messages_Tests: StreamTestCase {
 
     func test_deliveryStatusHidden_whenEphemeralMessageShown() throws {
         linkToScenario(withId: 439)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1324")
 
         GIVEN("user opens a channel") {
             userRobot
@@ -103,7 +96,7 @@ final class Ephemeral_Messages_Tests: StreamTestCase {
                 .openChannel()
         }
         WHEN("user runs a giphy command") {
-            userRobot.sendGiphy(send: false)
+            userRobot.uploadGiphy(send: false)
         }
         THEN("delivery status is hidden for ephemeral messages") {
             userRobot
@@ -114,17 +107,17 @@ final class Ephemeral_Messages_Tests: StreamTestCase {
 
     func test_deliveryStatusHidden_whenEphemeralMessageShownInThread() throws {
         linkToScenario(withId: 440)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1324")
 
         GIVEN("user opens a channel") {
-            backendRobot.generateChannels(count: 1, messagesCount: 1)
+            backendRobot.generateChannels(channelsCount: 1, messagesCount: 1)
             userRobot.login().openChannel()
         }
         WHEN("user runs a giphy command in thread") {
             userRobot
                 .openThread()
-                .sendGiphy(send: false)
+                .uploadGiphy(send: false)
         }
         THEN("delivery status is hidden for ephemeral messages") {
             userRobot
@@ -142,7 +135,7 @@ final class Ephemeral_Messages_Tests: StreamTestCase {
                 .openChannel()
         }
         WHEN("user sends a giphy using giphy command") {
-            userRobot.sendGiphy(useComposerCommand: true)
+            userRobot.uploadGiphy(useComposerCommand: true)
         }
         THEN("user observes the animated gif") {
             userRobot.assertGiphyImage()