+++ StreamChatSwiftUITestsAppTests/Tests/Message Delivery Status/MessageDeliveryStatus_Tests.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2025 Stream.io Inc. All rights reserved.
+// Copyright © 2026 Stream.io Inc. All rights reserved.
 //
 
 import XCTest
@@ -13,18 +13,10 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
     var pendingThreadReply: String { "pending \(threadReply)" }
     var failedThreadReply: String { "failed \(threadReply)" }
 
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        addTags([.messageDeliveryStatus])
-        assertMockServer()
-    }
-
     // MARK: Message List
 
     func test_singleCheckmarkShown_whenMessageIsSent() throws {
         linkToScenario(withId: 397)
-
-        throw XCTSkip("Check out SWUI-245")
 
         GIVEN("user opens chat") {
             userRobot
@@ -44,15 +36,13 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
     func test_deliveryStatusShowsClocks_whenMessageIsInPendingState() throws {
         linkToScenario(withId: 398)
 
-        throw XCTSkip("Check out SWUI-245")
-
         GIVEN("user opens the channel") {
             userRobot
                 .login()
                 .openChannel()
         }
         WHEN("user sends a new message") {
-            backendRobot.delayServerResponse(byTimeInterval: 5.0)
+            backendRobot.delayNewMessages(by: 5)
             userRobot.sendMessage(pendingMessage, waitForAppearance: false)
         }
         THEN("message delivery status shows clocks") {
@@ -62,8 +52,8 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
 
     func test_errorIndicatorShown_whenMessageFailedToBeSent() throws {
         linkToScenario(withId: 399)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1315")
 
         GIVEN("user becomes offline") {
             userRobot
@@ -88,8 +78,6 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
     func test_doubleCheckmarkShown_whenMessageReadByParticipant() throws {
         linkToScenario(withId: 400)
 
-        throw XCTSkip("Check out SWUI-245")
-
         GIVEN("user opens the channel") {
             userRobot
                 .login()
@@ -99,7 +87,7 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
             userRobot.sendMessage(message)
         }
         WHEN("participant reads the user's message") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
         }
         THEN("user spots double checkmark below the message") {
             userRobot.assertMessageDeliveryStatus(.read)
@@ -111,8 +99,8 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
 
     func test_doubleCheckmarkShown_whenNewParticipantAdded() throws {
         linkToScenario(withId: 401)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1316")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -123,7 +111,7 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
             userRobot.sendMessage(message)
         }
         AND("message is read by more than 1 participant") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
             userRobot
                 .assertMessageDeliveryStatus(.read)
                 .assertMessageReadCount(readBy: 1)
@@ -141,10 +129,8 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
 
     func test_readByDecremented_whenParticipantIsRemoved() throws {
         linkToScenario(withId: 402)
-
-        throw XCTSkip("Check out SWUI-245")
-
-        let participantOne = participantRobot.currentUserId
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1316")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -155,13 +141,13 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
             userRobot.sendMessage(message)
         }
         AND("is read by participant") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
             userRobot
                 .assertMessageDeliveryStatus(.read)
                 .assertMessageReadCount(readBy: 1)
         }
         WHEN("participant is removed from the channel") {
-//            userRobot.removeParticipant(withUserId: participantOne)
+//            userRobot.removeParticipant(withUserId: participantRobot.id)
         }
         THEN("user spots single checkmark below the message") {
             userRobot.assertMessageDeliveryStatus(.sent)
@@ -170,9 +156,8 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
 
     func test_deliveryStatusShownForTheLastMessageInGroup() throws {
         linkToScenario(withId: 403)
+        
         let secondMessage = "second message"
-
-        throw XCTSkip("Check out SWUI-245")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -198,8 +183,6 @@ final class MessageDeliveryStatus_Tests: StreamTestCase {
 
     func test_deliveryStatusHidden_whenMessageIsDeleted() throws {
         linkToScenario(withId: 404)
-
-        throw XCTSkip("Check out SWUI-245")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -231,8 +214,6 @@ extension MessageDeliveryStatus_Tests {
     func test_singleCheckmarkShown_whenMessageIsSent_andPreviewedInThread() throws {
         linkToScenario(withId: 405)
 
-        throw XCTSkip("Check out SWUI-245")
-
         GIVEN("user opens chat") {
             userRobot
                 .login()
@@ -253,8 +234,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_errorIndicatorShown_whenMessageFailedToBeSent_andCantBePreviewedInThread() throws {
         linkToScenario(withId: 406)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1315")
 
         GIVEN("user becomes offline") {
             userRobot
@@ -281,8 +262,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_doubleCheckmarkShown_whenMessageReadByParticipant_andPreviewedInThread() throws {
         linkToScenario(withId: 407)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-46")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -296,7 +277,7 @@ extension MessageDeliveryStatus_Tests {
             userRobot.openThread()
         }
         WHEN("the message is read by participant") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
         }
         THEN("user spots double checkmark below the message in thread") {
             userRobot.assertMessageDeliveryStatus(.read)
@@ -311,8 +292,6 @@ extension MessageDeliveryStatus_Tests {
     func test_singleCheckmarkShown_whenThreadReplyIsSent() throws {
         linkToScenario(withId: 408)
 
-        throw XCTSkip("Check out SWUI-245")
-
         GIVEN("user opens chat") {
             userRobot
                 .login()
@@ -322,7 +301,7 @@ extension MessageDeliveryStatus_Tests {
             participantRobot.sendMessage(message)
         }
         AND("user replies to the message in thread") {
-            userRobot.replyToMessageInThread(threadReply)
+            userRobot.sendMessageInThread(threadReply)
         }
         THEN("user spots single checkmark below the thread reply") {
             userRobot
@@ -333,11 +312,11 @@ extension MessageDeliveryStatus_Tests {
 
     func test_errorIndicatorShown_whenThreadReplyFailedToBeSent() throws {
         linkToScenario(withId: 409)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1315")
 
         GIVEN("user becomes offline") {
-            backendRobot.generateChannels(count: 1, messagesCount: 1)
+            backendRobot.generateChannels(channelsCount: 1, messagesCount: 1)
             userRobot
                 .setConnectivitySwitchVisibility(to: .on)
                 .login()
@@ -345,7 +324,7 @@ extension MessageDeliveryStatus_Tests {
                 .openChannel()
         }
         WHEN("user replies to message in thread") {
-            userRobot.replyToMessageInThread(failedThreadReply, waitForAppearance: false)
+            userRobot.sendMessageInThread(failedThreadReply, waitForAppearance: false)
         }
         THEN("error indicator is shown for the thread reply") {
             userRobot.assertThreadReplyFailedToBeSent()
@@ -359,8 +338,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_doubleCheckmarkShown_whenThreadReplyReadByParticipant() throws {
         linkToScenario(withId: 410)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-46")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -371,10 +350,10 @@ extension MessageDeliveryStatus_Tests {
             participantRobot.sendMessage(message)
         }
         WHEN("user replies to message in thread") {
-            userRobot.replyToMessageInThread(threadReply)
+            userRobot.sendMessageInThread(threadReply)
         }
         AND("participant reads the user's thread reply") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
         }
         THEN("user spots double checkmark below the message") {
             userRobot.assertMessageDeliveryStatus(.read)
@@ -386,8 +365,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_doubleCheckmarkShownInThreadReply_whenNewParticipantAdded() throws {
         linkToScenario(withId: 411)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1316")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -398,7 +377,7 @@ extension MessageDeliveryStatus_Tests {
             participantRobot.sendMessage(message)
         }
         AND("user replies to message in thread") {
-            userRobot.replyToMessageInThread(threadReply)
+            userRobot.sendMessageInThread(threadReply)
         }
         WHEN("new participant is added to the channel") {
 //            userRobot.addParticipant()
@@ -413,10 +392,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_readByDecrementedInThreadReply_whenParticipantIsRemoved() throws {
         linkToScenario(withId: 412)
-
-        throw XCTSkip("Check out SWUI-245")
-
-        let participantOne = participantRobot.currentUserId
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1316")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -427,16 +404,16 @@ extension MessageDeliveryStatus_Tests {
             participantRobot.sendMessage(message)
         }
         AND("user replies to message in thread") {
-            userRobot.replyToMessageInThread(threadReply)
+            userRobot.sendMessageInThread(threadReply)
         }
         AND("thread reply is read by participant") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
             userRobot
                 .assertMessageDeliveryStatus(.read)
                 .assertMessageReadCount(readBy: 1)
         }
         WHEN("participant is removed from the channel") {
-//            userRobot.removeParticipant(withUserId: participantOne)
+//            userRobot.removeParticipant(withUserId: participantRobot.id)
         }
         THEN("user spots single checkmark below the message") {
             userRobot.assertMessageDeliveryStatus(.sent)
@@ -445,8 +422,6 @@ extension MessageDeliveryStatus_Tests {
 
     func test_deliveryStatusShownForTheLastThreadReplyInGroup() throws {
         linkToScenario(withId: 413)
-
-        throw XCTSkip("Check out SWUI-245")
 
         let secondMessage = "second message"
 
@@ -459,7 +434,7 @@ extension MessageDeliveryStatus_Tests {
             participantRobot.sendMessage(message)
         }
         AND("user replies to message in thread") {
-            userRobot.replyToMessageInThread(threadReply)
+            userRobot.sendMessageInThread(threadReply)
         }
         AND("delivery status shows single checkmark") {
             userRobot.assertMessageDeliveryStatus(.sent)
@@ -478,8 +453,6 @@ extension MessageDeliveryStatus_Tests {
     func test_deliveryStatusHidden_whenThreadReplyIsDeleted() throws {
         linkToScenario(withId: 414)
 
-        throw XCTSkip("Check out SWUI-245")
-
         GIVEN("user opens the channel") {
             userRobot
                 .login()
@@ -489,7 +462,7 @@ extension MessageDeliveryStatus_Tests {
             participantRobot.sendMessage(message)
         }
         AND("user replies to message in thread") {
-            userRobot.replyToMessageInThread(threadReply)
+            userRobot.sendMessageInThread(threadReply)
         }
         AND("delivery status shows single checkmark") {
             userRobot.assertMessageDeliveryStatus(.sent)
@@ -506,8 +479,6 @@ extension MessageDeliveryStatus_Tests {
 
     func test_deliveryStatusShownForPreviousMessage_whenErrorMessageShown() throws {
         linkToScenario(withId: 415)
-
-        throw XCTSkip("Check out SWUI-245")
 
         GIVEN("user opens the channel") {
             userRobot
@@ -538,8 +509,6 @@ extension MessageDeliveryStatus_Tests {
     func test_deliveryStatusHidden_whenMessageIsSentAndReadEventsIsDisabled() throws {
         linkToScenario(withId: 416)
 
-        throw XCTSkip("Check out SWUI-245")
-
         GIVEN("user opens chat") {
             backendRobot.setReadEvents(to: false)
             userRobot
@@ -558,8 +527,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_deliveryStatusShowsClocks_whenMessageIsInPendingStateAndReadEventsIsDisabled() throws {
         linkToScenario(withId: 417)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-970")
 
         GIVEN("user opens the channel") {
             backendRobot.setReadEvents(to: false)
@@ -568,7 +537,7 @@ extension MessageDeliveryStatus_Tests {
                 .openChannel()
         }
         WHEN("user sends a new message") {
-            backendRobot.delayServerResponse(byTimeInterval: 5.0)
+            backendRobot.delayNewMessages(by: 5)
             userRobot.sendMessage(pendingMessage, waitForAppearance: false)
         }
         THEN("message delivery status shows clocks") {
@@ -580,8 +549,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_errorIndicatorShown_whenMessageFailedToBeSentAndReadEventsIsDisabled() throws {
         linkToScenario(withId: 418)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1315")
 
         GIVEN("user becomes offline") {
             backendRobot.setReadEvents(to: false)
@@ -607,8 +576,6 @@ extension MessageDeliveryStatus_Tests {
     func test_deliveryStatusHidden_whenMessageReadByParticipantAndReadEventsIsDisabled() throws {
         linkToScenario(withId: 419)
 
-        throw XCTSkip("Check out SWUI-245")
-
         GIVEN("user opens the channel") {
             backendRobot.setReadEvents(to: false)
             userRobot
@@ -619,7 +586,7 @@ extension MessageDeliveryStatus_Tests {
             userRobot.sendMessage(message)
         }
         WHEN("participant reads the user's message") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
         }
         THEN("delivery status is hidden") {
             userRobot
@@ -630,8 +597,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_deliveryStatusHidden_whenNewParticipantAddedAndReadEventsIsDisabled() throws {
         linkToScenario(withId: 420)
-
-        throw XCTSkip("Check out SWUI-245")
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1316")
 
         GIVEN("user opens the channel") {
             backendRobot.setReadEvents(to: false)
@@ -643,7 +610,7 @@ extension MessageDeliveryStatus_Tests {
             userRobot.sendMessage(message)
         }
         AND("message is read by more than 1 participant") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
         }
         WHEN("new participant is added to the channel") {
 //            userRobot.addParticipant()
@@ -657,10 +624,8 @@ extension MessageDeliveryStatus_Tests {
 
     func test_deliveryStatusHidden_whenParticipantIsRemovedAndReadEventsIsDisabled() throws {
         linkToScenario(withId: 421)
-
-        throw XCTSkip("Check out SWUI-245")
-
-        let participantOne = participantRobot.currentUserId
+        
+        throw XCTSkip("https://linear.app/stream/issue/IOS-1316")
 
         GIVEN("user opens the channel") {
             backendRobot.setReadEvents(to: false)
@@ -672,10 +637,10 @@ extension MessageDeliveryStatus_Tests {
             userRobot.sendMessage(message)
         }
         AND("is read by participant") {
-            participantRobot.readMessageAfterDelay()
+            participantRobot.readMessage()
         }
         WHEN("participant is removed from the channel") {
-//            userRobot.removeParticipant(withUserId: participantOne)
+//            userRobot.removeParticipant(withUserId: participantRobot.id)
         }
         AND("delivery status is hidden") {
             userRobot.assertMessageDeliveryStatus(nil)
@@ -687,8 +652,6 @@ extension MessageDeliveryStatus_Tests {
 
     func test_deliveryStatusHiddenForMessagesInGroup_whenReadEventsIsDisabled() throws {
         linkToScenario(withId: 422)
-
-        throw XCTSkip("Check out SWUI-245")
 
         let secondMessage = "second message"
 
@@ -717,8 +680,6 @@ extension MessageDeliveryStatus_Tests {
 
     func test_deliveryStatusHidden_whenMessageIsDeletedAndReadEventsIsDisabled() throws {
         linkToScenario(withId: 423)
-
-        throw XCTSkip("Check out SWUI-245")
 
         GIVEN("user opens the channel") {
             backendRobot.setReadEvents(to: false)